Higher confidence changes
