+++ src/commands/pruneSystem.ts
@@ -32,14 +32,14 @@ export async function pruneSystem(context: IActionContext): Promise<void> {
 
             let message: string;
             if (containersResult?.containersDeleted?.length && Number.isInteger(containersResult?.spaceReclaimed) &&
-                imagesResult?.imagesDeleted?.length && Number.isInteger(imagesResult?.spaceReclaimed) &&
+                imagesResult?.imageRefsDeleted?.length && Number.isInteger(imagesResult?.spaceReclaimed) &&
                 networksResult?.networksDeleted?.length &&
                 volumesResult?.volumesDeleted?.length && Number.isInteger(volumesResult?.spaceReclaimed)) {
                 message = localize(
                     'vscode-docker.commands.pruneSystem.removed',
                     'Removed {0} container(s), {1} image(s), {2} network(s), {3} volume(s) and reclaimed {4} MB of space.',
                     containersResult.containersDeleted.length,
-                    imagesResult.imagesDeleted.length,
+                    imagesResult.imageRefsDeleted.length,
                     networksResult.networksDeleted.length,
                     volumesResult.volumesDeleted.length,
                     convertToMB(containersResult.spaceReclaimed + imagesResult.spaceReclaimed + volumesResult.spaceReclaimed)
