+++ src/tree/containers/ContainerGroupTreeItem.ts
@@ -8,8 +8,8 @@ import { ThemeIcon, TreeItemCollapsibleState } from "vscode";
 import { LocalGroupTreeItemBase } from "../LocalGroupTreeItemBase";
 import { LocalRootTreeItemBase } from "../LocalRootTreeItemBase";
 import { getCommonGroupIcon } from "../settings/CommonProperties";
-import { ContainerProperty, getContainerStateIcon } from "./ContainerProperties";
-import { DockerContainerInfo, NonComposeGroupName } from "./ContainersTreeItem";
+import { ContainerProperty, getContainerStateIcon, NonComposeGroupName } from "./ContainerProperties";
+import { DockerContainerInfo } from "./ContainersTreeItem";
 
 export class ContainerGroupTreeItem extends LocalGroupTreeItemBase<DockerContainerInfo, ContainerProperty> {
     public childTypeLabel: string = 'container';
