+++ src/tree/containers/ContainerProperties.ts
@@ -3,9 +3,10 @@
  *  Licensed under the MIT License. See LICENSE.md in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { ListContainersItem } from "@microsoft/container-runtimes";
 import { ThemeColor, ThemeIcon } from "vscode";
 import { localize } from "../../localize";
-import { commonProperties, CommonProperty } from "../settings/CommonProperties";
+import { commonProperties, CommonProperty, getCommonPropertyValue } from "../settings/CommonProperties";
 import { ITreePropertyInfo } from "../settings/ITreeSettingInfo";
 
 export type ContainerProperty = Exclude<CommonProperty, 'Size'> | 'Image' | 'Compose Project Name' | 'ContainerId' | 'ContainerName' | 'Networks' | 'Ports' | 'State' | 'Status';
@@ -39,5 +40,51 @@ export function getContainerStateIcon(state: string): ThemeIcon {
         case 'running':
         default:
             return new ThemeIcon('debug-start', new ThemeColor('debugIcon.startForeground'));
+    }
+}
+
+export function getContainerPropertyValue(item: ListContainersItem, property: ContainerProperty): string {
+    const networks = item.networks?.length > 0 ? item.networks : ['<none>'];
+    const ports = item.ports?.length > 0 ? item.ports.map(p => p.hostPort) : ['<none>'];
+
+    switch (property) {
+        case 'ContainerId':
+            return item.id.slice(0, 12);
+        case 'ContainerName':
+            return item.name;
+        case 'Networks':
+            return networks.join(',');
+        case 'Ports':
+            return ports.join(',');
+        case 'State':
+            return item.state;
+        case 'Status':
+            // The rapidly-refreshing status during a container's first minute causes a lot of problems with excessive refreshing
+            // This normalizes things like "10 seconds" to "Less than a minute", meaning the refreshes don't happen constantly
+            return item.status?.replace(/\d+ seconds?/i, localize('vscode-docker.tree.containers.lessThanMinute', 'Less than a minute'));
+        case 'Compose Project Name':
+            return getComposeProjectName(item);
+        case 'Image':
+            return item.image.originalName;
+        default:
+            return getCommonPropertyValue(item, property);
+    }
+}
+
+export const NonComposeGroupName = localize('vscode-docker.tree.containers.otherContainers', 'Individual Containers');
+
+export function getComposeProjectName(container: ListContainersItem): string {
+    if (!container.labels) {
+        return NonComposeGroupName;
+    }
+
+    const labels = Object.keys(container.labels)
+        .map(label => ({ label: label, value: container.labels[label] }));
+
+    const composeProject = labels.find(l => l.label === 'com.docker.compose.project');
+    if (composeProject) {
+        return composeProject.value;
+    } else {
+        return NonComposeGroupName;
     }
 }
